Add backward-compatible connect() method for PyQt5
In PyQt5, QApplication no longer has a connect() method for old-style
SIGNAL/SLOT connections. This was causing AttributeError when code tried
to use self.connect().

Added a connect() compatibility method that:
1. Accepts old PyQt4-style: self.connect(sender, SIGNAL('name'), slot)
2. Parses the SIGNAL string to extract the signal name
3. Converts to new PyQt5-style: sender.signal_name.connect(slot)
4. Provides helpful warnings if signals can't be found

This allows all 163 old-style signal connections in the codebase to work
without modification. A proper migration to new-style signals can be done
later, but this gets the app working now.